reactivated unit test
